backend/product_service/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
python-decouple==3.8
drf-yasg==1.21.10
requests==2.31.0
Pillow==10.1.0
django-filter==25.1
